requirements.txt
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
emoji==2.14.0
filelock==3.16.1
fsspec==2024.10.0
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.1
nltk==3.9.1
numpy==1.24.4
pillow==10.4.0
protobuf==5.28.3
PyPDF2==3.0.1
python-docx==1.1.2
regex==2024.11.6
requests==2.32.3
stanza==1.9.2
sympy==1.13.3
tika==2.6.0
tomli==2.1.0
torch==2.4.1
tqdm==4.67.0
typing_extensions==4.12.2
Unidecode==1.3.8
urllib3==2.2.3
Whoosh==2.7.4
